[PoC] Fix api path and fields

+++ core/network/src/commonMain/kotlin/com/dezdeqness/network/services/CatalogService.kt
@@ -17,6 +17,6 @@ interface CatalogService {
     ): Response<GeneralResponse<List<ReleaseResponse>>>
 
     @GET(RELEASE)
-    suspend fun getReleaseById(@Path id: Int): Response<ReleaseResponse>
+    suspend fun getReleaseById(@Path id: Long): Response<ReleaseResponse>
 
 }
