+++ features/details/src/commonMain/kotlin/com/dezdeqness/details/domain/repository/ReleaseRepository.kt
@@ -4,6 +4,6 @@ import com.dezdeqness.details.domain.model.ReleaseDetailsEntity
 
 interface ReleaseRepository {
 
-    suspend fun getReleaseById(id: Int): Result<ReleaseDetailsEntity>
+    suspend fun getReleaseById(id: Long): Result<ReleaseDetailsEntity>
 
 }
